change name to include backstop apr and add direction prop
